refactor: move `safeTruncate` from `github.ts` to `utils.ts`
